// src/types/supabase.ts
export type Json =
    | string
    | number
    | boolean
    | null
    | { [key: string]: Json | undefined }
    | Json[]

export type Database = {
    // Allows to automatically instantiate createClient with right options
    // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
    __InternalSupabase: {
        PostgrestVersion: "14.1"
    }
    public: {
        Tables: {
// ...
            experience: {
                Row: {
                    achievements: Json | null
                    created_at: string | null
                    description: string | null
                    display_order: number | null
                    id: string
                    language: string
                    location: string
                    organization: string
                    period: string
                    title: string
                }
                Insert: {
                    achievements?: Json | null
                    created_at?: string | null
                    description?: string | null
                    display_order?: number | null
                    id?: string
                    language?: string
                    location: string
                    organization: string
                    period: string
                    title: string
                }
                Update: {
                    achievements?: Json | null
                    created_at?: string | null
                    description?: string | null
                    display_order?: number | null
                    id?: string
                    language?: string
                    location?: string
                    organization?: string
                    period?: string
                    title?: string
                }
                Relationships: []
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}